[ISSUES#] 2739 provider health check for sozo.
