+++ bin/sozo/src/commands/mod.rs
@@ -34,6 +34,8 @@ use migrate::MigrateArgs;
 use model::ModelArgs;
 use test::TestArgs;
 
+pub(crate) const LOG_TARGET: &str = "sozo::cli";
+
 #[derive(Debug, Subcommand)]
 pub enum Commands {
     #[command(about = "Grant or revoke a contract permission to write to a resource")]
@@ -42,8 +44,10 @@ pub enum Commands {
     Build(Box<BuildArgs>),
     #[command(about = "Build and migrate the world every time a file changes")]
     Dev(Box<DevArgs>),
-    #[command(about = "Run a migration, declaring and deploying contracts as necessary to update \
-                       the world")]
+    #[command(
+        about = "Run a migration, declaring and deploying contracts as necessary to update \
+                       the world"
+    )]
     Migrate(Box<MigrateArgs>),
     #[command(about = "Execute a system with the given calldata.")]
     Execute(Box<ExecuteArgs>),
