+++ crates/dojo/utils/src/lib.rs
@@ -14,3 +14,4 @@ pub mod env;
 pub mod keystore;
 
 pub mod signal;
+pub mod utils;
