+++ crates/torii/core/Cargo.toml
@@ -18,6 +18,7 @@ chrono.workspace = true
 crypto-bigint.workspace = true
 data-url.workspace = true
 dojo-types.workspace = true
+dojo-utils.workspace = true
 dojo-world.workspace = true
 futures-channel = "0.3.0"
 futures-util.workspace = true
